Added log in WS
