--- studio-api/lib/live-srt-lib/README.md
@@ -1,3 +1,0 @@
-# Lib component
-
-This component is the shared library used in all other components. It contains the database model in model/model.js as well as the basic building blocks, namely the application, the component, the MQTT client, and the configuration management.
