--- studio-api/lib/live-srt-lib/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  Component: require("./component.js"),
-  MqttClient: require("./mqtt.js"),
-  CustomErrors: require("./customErrors.js"),
-}
